storage: Move sqlite errors handling to shared file
