+++ src/storage/sqlite_common.rs
@@ -3,10 +3,60 @@
 
 use std::sync::MutexGuard;
 
-use crate::error::Result;
+use crate::error::{Error, Result};
 use crate::interface::*;
+use rusqlite::{Error as rlError, ErrorCode};
 
 use rusqlite;
+
+impl From<rlError> for Error {
+    fn from(error: rlError) -> Error {
+        match error {
+            rlError::SqliteFailure(_, _) => match error.sqlite_error_code() {
+                Some(e) => match e {
+                    ErrorCode::ConstraintViolation
+                    | ErrorCode::TypeMismatch
+                    | ErrorCode::ApiMisuse
+                    | ErrorCode::ParameterOutOfRange => {
+                        Error::ck_rv_from_error(CKR_GENERAL_ERROR, error)
+                    }
+                    ErrorCode::DatabaseBusy
+                    | ErrorCode::DatabaseLocked
+                    | ErrorCode::FileLockingProtocolFailed => {
+                        Error::ck_rv_from_error(
+                            CKR_TOKEN_RESOURCE_EXCEEDED,
+                            error,
+                        )
+                    }
+                    ErrorCode::OutOfMemory => {
+                        Error::ck_rv_from_error(CKR_DEVICE_MEMORY, error)
+                    }
+                    ErrorCode::CannotOpen
+                    | ErrorCode::NotFound
+                    | ErrorCode::PermissionDenied => {
+                        Error::ck_rv_from_error(CKR_TOKEN_NOT_RECOGNIZED, error)
+                    }
+                    ErrorCode::ReadOnly => Error::ck_rv_from_error(
+                        CKR_TOKEN_WRITE_PROTECTED,
+                        error,
+                    ),
+                    ErrorCode::TooBig => {
+                        Error::ck_rv_from_error(CKR_DATA_LEN_RANGE, error)
+                    }
+                    _ => Error::ck_rv_from_error(CKR_DEVICE_ERROR, error),
+                },
+                None => Error::ck_rv_from_error(CKR_GENERAL_ERROR, error),
+            },
+            _ => Error::ck_rv_from_error(CKR_GENERAL_ERROR, error),
+        }
+    }
+}
+
+impl<T> From<std::sync::PoisonError<std::sync::MutexGuard<'_, T>>> for Error {
+    fn from(_: std::sync::PoisonError<std::sync::MutexGuard<'_, T>>) -> Error {
+        Error::ck_rv(CKR_CANT_LOCK)
+    }
+}
 
 pub fn check_table(
     conn: &MutexGuard<'_, rusqlite::Connection>,
